move non-crud-controllers out of 'admin'
