+++ app/Http/Controllers/MyRoleController.php
@@ -1,14 +1,13 @@
 <?php
 
-namespace App\Http\Controllers\Admin;
+namespace App\Http\Controllers;
 
-use Carbon\Carbon;
-use App\Models\User;
-use Illuminate\Http\Request;
-use App\Models\RemovalRequest;
 use App\Mail\DataRemovalRequested;
 use App\Models\OrganisationMember;
-use App\Http\Controllers\Controller;
+use App\Models\RemovalRequest;
+use App\Models\User;
+use Carbon\Carbon;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Session;
@@ -53,7 +52,7 @@ class MyRoleController extends Controller
             $user->save();
         }
 
-        // logout user 
+        // logout user
         Auth::logout();
 
         // redirect user to login page
@@ -93,6 +92,6 @@ class MyRoleController extends Controller
             'organisation' => $organisation,
         ]);
 
-    }    
+    }
 
 }
