+++ routes/backpack/custom.php
@@ -13,24 +13,23 @@ use App\Http\Controllers\Admin\ContinentCrudController;
 use App\Http\Controllers\Admin\CountryCrudController;
 use App\Http\Controllers\Admin\OrganisationCrudController;
 use App\Http\Controllers\Admin\PortfolioCrudController;
-use App\Http\Controllers\Admin\PrincipleAssessmentCrudController;
 use App\Http\Controllers\Admin\PrincipleCrudController;
 use App\Http\Controllers\Admin\ProjectCrudController;
 use App\Http\Controllers\Admin\RedLineCrudController;
 use App\Http\Controllers\Admin\RegionCrudController;
+use App\Http\Controllers\Admin\RemovalRequestCrudController;
 use App\Http\Controllers\Admin\RoleInviteCrudController;
 use App\Http\Controllers\Admin\ScoreTagCrudController;
 use App\Http\Controllers\Admin\UserCrudController;
-use App\Http\Controllers\Admin\SelectedOrganisationController;
-use App\Http\Controllers\Admin\DashboardController;
-use App\Http\Controllers\Admin\GenericDashboardController;
-use App\Http\Controllers\Admin\MyRoleController;
-use App\Http\Controllers\Admin\RemovalRequestCrudController;
 use App\Http\Controllers\AssessmentController;
+use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\GeneratePdfFileController;
+use App\Http\Controllers\GenericDashboardController;
+use App\Http\Controllers\MyRoleController;
 use App\Http\Controllers\OrganisationController;
 use App\Http\Controllers\OrganisationMemberController;
-use App\Http\Controllers\GeneratePdfFileController;
 use App\Http\Controllers\ProjectController;
+use App\Http\Controllers\SelectedOrganisationController;
 
 Route::group([
     'prefix' => config('backpack.base.route_prefix', 'admin'),
